Use freed up name for standard root readers / writers
